+++ .github/workflows/ci-build.yml
@@ -52,6 +52,8 @@ jobs:
       uses: Swatinem/rust-cache@v2
     - name: Build
       run: cargo test --no-run --locked
+    - name: Build minimal janus_messages
+      run: cargo build --package janus_messages --no-default-features
     - name: Test
       id: test
       env:
