+++ .github/workflows/publish-crates.yml
@@ -13,9 +13,15 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v3
-    # We retry crate publishing to allow the newly published janus_core to be
-    # visible on crates.io when we get around to publishing janus_client and
-    # janus_collector.
+    # We retry crate publishing to allow the newly published janus_messages to be visible on
+    # crates.io when we get around to publishing janus_core, janus_client and janus_collector.
+    - name: "Publish janus_messages"
+      uses: nick-fields/retry@v2
+      with:
+        timeout_minutes: 10
+        max_attempts: 5
+        polling_interval_seconds: 30
+        command: cargo publish -p janus_messages
     - name: "Publish janus_core"
       uses: nick-fields/retry@v2
       with:
