+++ Dockerfile
@@ -12,6 +12,7 @@ COPY interop_binaries /src/interop_binaries
 COPY janus_collector /src/janus_collector
 COPY janus_core /src/janus_core
 COPY janus_client /src/janus_client
+COPY janus_messages /src/janus_messages
 COPY janus_server /src/janus_server
 RUN --mount=type=cache,target=/usr/local/cargo/registry --mount=type=cache,target=/src/target cargo build --release -p janus_server --bin $BINARY --features=prometheus && cp /src/target/release/$BINARY /$BINARY
