+++ Dockerfile.interop
@@ -15,6 +15,7 @@ COPY interop_binaries /src/interop_binaries
 COPY janus_collector /src/janus_collector
 COPY janus_core /src/janus_core
 COPY janus_client /src/janus_client
+COPY janus_messages /src/janus_messages
 COPY janus_server /src/janus_server
 RUN --mount=type=cache,target=/usr/local/cargo/registry --mount=type=cache,target=/src/target cargo build --profile $PROFILE -p interop_binaries --bin $BINARY && cp /src/target/$PROFILE/$BINARY /$BINARY
