+++ integration_tests/Cargo.toml
@@ -25,6 +25,7 @@ futures = "0.3.24"
 hex = "0.4"
 interop_binaries = { path = "../interop_binaries", features = ["testcontainer"] }
 janus_core = { path = "../janus_core", features = ["test-util"] }
+janus_messages = { path = "../janus_messages" }
 janus_server = { path = "../janus_server", features = ["test-util"] }
 k8s-openapi = { version = "0.13.1", features = ["v1_20"] }  # keep this version in sync with what is referenced by the indirect dependency via `kube`
 kube = { version = "0.65.0", default-features = false, features = ["client"] }
