+++ janus_collector/src/lib.rs
@@ -8,9 +8,10 @@
 //!
 //! ```no_run
 //! use janus_collector::{Collector, CollectorParameters, default_http_client};
-//! use janus_core::{
-//!     message::{Duration, HpkeConfig, Interval, TaskId, Time},
-//!     task::AuthenticationToken,
+//! use janus_core::{hpke::generate_hpke_config_and_private_key, task::AuthenticationToken};
+//! use janus_messages::{
+//!     Duration, HpkeAeadId, HpkeConfig, HpkeConfigId, HpkeKdfId, HpkeKemId, Interval, TaskId,
+//!     Time,
 //! };
 //! use prio::vdaf::prio3::Prio3;
 //! use rand::random;
@@ -18,13 +19,13 @@
 //!
 //! # async fn run() {
 //! // Supply DAP task paramenters.
-//! # let task_id = random();
-//! # let (hpke_config, private_key) = janus_core::hpke::generate_hpke_config_and_private_key(
-//! #     janus_core::message::HpkeConfigId::from(0),
-//! #     janus_core::message::HpkeKemId::X25519HkdfSha256,
-//! #     janus_core::message::HpkeKdfId::HkdfSha256,
-//! #     janus_core::message::HpkeAeadId::Aes128Gcm,
-//! # );
+//! let task_id = random();
+//! let (hpke_config, private_key) = janus_core::hpke::generate_hpke_config_and_private_key(
+//!     HpkeConfigId::from(0),
+//!     HpkeKemId::X25519HkdfSha256,
+//!     HpkeKdfId::HkdfSha256,
+//!     HpkeAeadId::Aes128Gcm,
+//! );
 //! let authentication_token = AuthenticationToken::from(b"my-authentication-token".to_vec());
 //! let parameters = CollectorParameters::new(
 //!     task_id,
@@ -55,12 +56,11 @@ use backoff::{backoff::Backoff, ExponentialBackoff};
 use derivative::Derivative;
 use janus_core::{
     hpke::{self, associated_data_for_aggregate_share, HpkeApplicationInfo, HpkePrivateKey},
-    message::{
-        query_type::TimeInterval, CollectReq, CollectResp, HpkeConfig, Interval, Query, Role,
-        TaskId,
-    },
     retries::{http_request_exponential_backoff, retry_http_request},
     task::{url_ensure_trailing_slash, AuthenticationToken, DAP_AUTH_HEADER},
+};
+use janus_messages::{
+    query_type::TimeInterval, CollectReq, CollectResp, HpkeConfig, Interval, Query, Role, TaskId,
 };
 use prio::{
     codec::{Decode, Encode},
@@ -483,7 +483,7 @@ where
 #[cfg(feature = "test-util")]
 pub mod test_util {
     use crate::{Collector, Error};
-    use janus_core::message::Interval;
+    use janus_messages::Interval;
     use prio::vdaf;
 
     pub async fn collect_with_rewritten_url<V: vdaf::Collector>(
@@ -512,10 +512,10 @@ mod tests {
     use chrono::{TimeZone, Utc};
     use janus_core::{
         hpke::{test_util::generate_test_hpke_config_and_private_key, Label},
-        message::{Duration, HpkeCiphertext, PartialBatchSelector, Time},
         retries::test_http_request_exponential_backoff,
         test_util::{install_test_trace_subscriber, run_vdaf, VdafTranscript},
     };
+    use janus_messages::{Duration, HpkeCiphertext, PartialBatchSelector, Time};
     use mockito::mock;
     use prio::{
         field::Field64,
