+++ janus_messages/README.md
@@ -1,0 +1,3 @@
+# `janus_messages`
+
+This crate provides structures representing the various messages defined by the [Distributed Aggregation Protocol](https://datatracker.ietf.org/doc/draft-ietf-ppm-dap/).
