+++ janus_server/Cargo.toml
@@ -10,9 +10,19 @@ rust-version = "1.63"
 default = ["kube-rustls"]
 tokio-console = ["dep:console-subscriber"]
 jaeger = ["dep:tracing-opentelemetry", "dep:opentelemetry-jaeger"]
-otlp = ["dep:tracing-opentelemetry", "dep:opentelemetry-otlp", "dep:opentelemetry-semantic-conventions", "dep:tonic"]
+otlp = [
+    "dep:tracing-opentelemetry",
+    "dep:opentelemetry-otlp",
+    "dep:opentelemetry-semantic-conventions",
+    "dep:tonic",
+]
 prometheus = ["dep:opentelemetry-prometheus", "dep:prometheus"]
-test-util = ["janus_core/test-util", "dep:lazy_static", "dep:testcontainers"]
+test-util = [
+    "janus_core/test-util",
+    "janus_messages/test-util",
+    "dep:lazy_static",
+    "dep:testcontainers",
+]
 kube-rustls = ["kube/rustls-tls"]
 kube-openssl = ["kube/openssl-tls"]
 
@@ -29,21 +39,20 @@ deadpool = { version = "0.9.5", features = ["rt_tokio_1"] }
 deadpool-postgres = "0.10.1"
 derivative = "2.2.0"
 futures = "0.3.24"
-hex = "0.4.3"
-hpke-dispatch = "0.3.0"
 http = "0.2.8"
 hyper = "0.14.20"
 itertools = "0.10.5"
-janus_core = { path = "../janus_core", features = ["database"] }
+janus_core = { path = "../janus_core" }
+janus_messages = { path = "../janus_messages" }
 k8s-openapi = { version = "0.13.1", features = ["v1_20"] }  # keep this version in sync with what is referenced by the indirect dependency via `kube`
 kube = { version = "0.65.0", default-features = false, features = ["client"] }
 lazy_static = { version = "1", optional = true }
-num_enum = "0.5.6"
 opentelemetry = { version = "0.17.0", features = ["metrics", "rt-tokio"] }
 opentelemetry-jaeger = { version = "0.16.0", optional = true, features = ["rt-tokio"] }
 opentelemetry-otlp = { version = "0.10.0", optional = true, features = ["metrics"] }  # ensure that the version of tonic below matches what this uses
 opentelemetry-prometheus = { version = "0.10.0", optional = true }
 opentelemetry-semantic-conventions = { version = "0.9.0", optional = true }
+postgres-protocol = "0.6.4"
 postgres-types = { version = "0.2.4", features = ["derive", "array-impls"] }
 prio = { version = "0.8.2", features = ["multithreaded"] }
 prometheus = { version = "0.13.2", optional = true }
