+++ janus_server/src/bin/collect_job_driver.rs
@@ -1,6 +1,7 @@
 use anyhow::Context;
 use clap::Parser;
-use janus_core::{message::Duration, time::RealClock, TokioRuntime};
+use janus_core::{time::RealClock, TokioRuntime};
+use janus_messages::Duration;
 use janus_server::{
     aggregator::aggregate_share::CollectJobDriver,
     binary_utils::{
